Refactor the cfg_firmware_web.rb script
a. Add a method read_ip_addr to the script to read ip address from hosts
file. The reason is that it need to conver from test_site to ip address
to avoid the problem of across-domain visit when uploading firmware.
b. In the ensure section, it should convert back from ip address to
test_site and go back to configure page for the next script can run.
c. It will get stuck when clicking firmware->web link or go back to
configure tab from firmware->web page because of waiting. The changes at
the login part that login equipinfo first then click updtweb link is a
solution for this issue. See DEF3892 for the detail.
d. The script cannot run on chinese OS. See DEF3897 for detail reason.
e. Change the definition of 'web_updt' button in navigate file.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0