After reading table info, it should go back to configure page
1) It will get stuck when clicking firmware->web link or go back to
configure tab from firmware->web page because of waiting. The changes
that let it login to equipinfo first then go to firmware->web link is
a solution for this issue for now.
2) In ensure section, let it go back to agent info page and handle
the popup,so next scripts can run from that point.
Please also see DEF3892 in Monitoring Test-Firmware project for this issue.

IS-WEBCARD_HID9_5.300.0_065740; Watir 1.7.0