Revert "Change several DNS testcases in the spreadsheet to fit changes on the card"
This change was made before the discussion of WB1.0,WB2.0,WB3.0 and WB4.0.
This change works on IPv6 card(WB2.0), but not work on SNMPv3 card(WB1.0).
Since the WB1.0 has NOT been taged yet, revert this change to let it work
on SNMPv3 card(WB1.0) first.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0

This reverts commit ee4371f94732787083590ef5c60bb07ced507a6d.