Revert "Revert "Change several DNS testcases in the spreadsheet to fit changes"
Currently, this is in Watir2.0, so the very orinigal change "Change several DNS
testcases in the spreadsheet to fit changes in the card" is necessary now.
The change of the card of DNS page is that, when input invalid ip addresses, it
does NOT reject quietly any more but pop up a warning box now.

IS-WEBCARD_HID9_5.300.0_066280; Watir 1.7.0

This reverts commit 64e164936955e9eb5a275d370d346ed3d7afab2c.