Correct the last step's expected values in the spreadsheet
The last step of the script is to restore the page to default status.
And, the original expected value of 'Server'is '0.0.0.0','Filename'
is 'Uninitialized'. However, this is not an invalid filename for lack
of '.bin' extension.
So, in the last step, when input '0.0.0.0' and 'Uninitialized', it should
be saved as the previous valid value.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0