The script does NOT read all values in the web page
When read default values from web and write to the spreadsheet, script
does NOT read all the values on the web page, and one value write to the
wrong cell of the spreadsheet.
Fix it by copying the corresponding scripts from cfg_snmp_wb.rb to replace
the old ones.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0