Recursively return a dir list of test cases from Testlog
The specific DIRECTORY path is determined in "dir = []. All test cases
in that directory and sub-directories below will be returned along with
the Title of the test case "test_title"

The Testlog ".tlg" files are .xml format.  This script can be modified
to extract any information desired by specifying the desired xml xpath.