Implement REQ6285-Time stamped test result sub-folder
1)In controller script
a. Create the result sub-folder under the result folder.
b. Pass the result sub-folder name to the setup method of controller.
c. Pass the result sub-folder name as an ARGV to the drivers when
run from controller.
2)In setup method
a. Pass the sub-folder name to the method xls_timestamp.
3)In xls_timestamp method
a. Add the sub-folder name to the new path of save-as spreadsheet.

IS-WEBCARD_HID9_5.300.0_067239; Watir 1.7.0