Refactor web_controller_telnet
Most of the original web_controller_telnet script methods are duplicate
with the ones in telnet_prototype.rb.
1)System call telnet_prototype.rb to execute each checked script in the
controller spreadsheet.
2)Change the path in controller spreadsheet from /telnet/*** to ***. The
reason is that, when run individually from telnet_prototype.rb, It just
need to input the driver script name. So it's better to pass the same thing
when system call it.
3)When run from controller, the results will be saved into a sub-folder
of result folder, but when run from telnet_prototype.rb, the result will
be saved into result folder. This is consistent with Watir scripts.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0