Move the cell of test status of telnet drivers from B16 to B9
1) Watir driver scripts judge pass/fail in B9 cell. And the tear_down_d
method get the run status from B9 cell.So if telnet scripts want to call
tear_down_d method to update controller spreadsheet, the drivers
spreadsheet should be consistent with the Watir scripts spreadsheet,
that use B9 cell to contain execution result, instead of B16.
2) Also, it is not very good that,the cells containing card info,
IE version and ruby/watir version are ahead of result cell in spreadsheet.
Because,it is not convience when we want to collect more information
when setup execution in the future.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0