Optimize the select_test_suite method in controller.rb
Do NOT use hash to contain the controller spreadsheets, but use an Array.
Because when use the hash, and add a new entry (web_controller_telnet.xls)
it will display not in order, like below
The following test suites are available for execution:
5 - web_controller_telnet
1 - controller_rpc
2 - controller_web
3 - controller_webx
4 - controller_web_ipv6
Please type the number of the desired suite followed by <Enter>
But the expectation is
The following test suites are available for execution:
1 - controller_rpc
2 - controller_web
3 - controller_webx
4 - controller_web_ipv6
5 - web_controller_telnet
Please type the number of the desired suite followed by <Enter>
So, this change is needed.

IS-WEBCARD_HID9_5.200.0_059101; Watir 1.7.0