Modify datamapping scripts - Add local methods
Because the definition of supprt in navigate.rb are not valid for
IS-WEBL card. So make the definition locally. In this way, the
methods that calling supprt will have to be defined locally also
in serial.
This commit just copy those methods to this scirpt and make little
changes to let them work. While, next commit will clear up the methods
to let them looks shorter.
Also, change the spreadsheet the test site address from a specific
IP address to 'test_site'

IS-WEBL_HID10_5.280.0_069012;Watir 1.7.0