Add Tools\temp_files folder to hold general input and output files
This is where work files will be manipulated.  An example would be
processing a LIFE.net mapping file to create testlog test cases.
APM_Life mapping.xlsx would be placed in this folder.  The script
would then process the file and place the output file - xxx.csv
to be used for import into Testlog.